move AvroIO/taps into com.spotify.scio.avro package (#1446)
