+++ scio-examples/src/test/scala/com/spotify/scio/examples/extra/AvroInOutTest.scala
@@ -18,7 +18,6 @@
 package com.spotify.scio.examples.extra
 
 import com.spotify.scio.avro._
-import com.spotify.scio.io._
 import com.spotify.scio.testing._
 
 import scala.collection.JavaConverters._
