+++ scio-examples/src/test/scala/com/spotify/scio/examples/extra/ProtobufExampleTest.scala
@@ -17,6 +17,7 @@
 
 package com.spotify.scio.examples.extra
 
+import com.spotify.scio.avro.ProtobufIO
 import com.spotify.scio.proto.SimpleV2.SimplePB
 import com.spotify.scio.proto.Track.TrackPB
 import com.spotify.scio.io._
