+++ scio-examples/src/test/scala/com/spotify/scio/examples/extra/ShapelessAvroExampleTest.scala
@@ -17,6 +17,7 @@
 
 package com.spotify.scio.examples.extra
 
+import com.spotify.scio.avro.AvroIO
 import com.spotify.scio.io._
 import com.spotify.scio.testing._
 import org.apache.avro.generic.{GenericData, GenericRecord}
